smooth only once (move the final smoothing before drivable area calc)
